Split the settings search catalog under the file size limit.
