+++ app/scripts/comp/settings/settings-search-item.js
@@ -1,0 +1,5 @@
+function item(id, page, section, target, titleKey, categoryKey, keywords = [], visible) {
+    return { id, page, section, target, titleKey, categoryKey, keywords, visible };
+}
+
+export { item };
